requirements.txt
beautifulsoup4==4.7.1
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.1
django-htmlmin==0.11.0
django-recaptcha2==1.4.1
django-solo==1.1.3
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
mistune==0.8.4
mysqlclient==1.4.2.post1
Pillow==6.0.0
psycopg2==2.8.2
Pygments==2.4.2
python-dotenv==0.10.2
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
urllib3==1.25.3
webencodings==0.5.1
whitenoise==4.1.2
